20.0.650

+++ modules/im/install/js/im/provider/pull/src/im.command.handler.js
@@ -9,6 +9,7 @@
 
 import {PullClient} from "pull.client";
 import {VuexBuilderModel} from 'ui.vue.vuex';
+import {EventType} from "im.const";
 
 class ImPullCommandHandler
 {
@@ -27,6 +28,20 @@ class ImPullCommandHandler
 		{
 			this.store = params.store;
 		}
+
+		this.option = typeof params.store === 'object' && params.store? params.store: {};
+
+		if (
+			!(
+				typeof this.option.handlingDialog === 'object'
+				&& this.option.handlingDialog
+				&& this.option.handlingDialog.chatId
+				&& this.option.handlingDialog.dialogId
+			)
+		)
+		{
+			this.option.handlingDialog = false;
+		}
 	}
 
 	getModuleId()
@@ -39,12 +54,67 @@ class ImPullCommandHandler
 		return PullClient.SubscriptionType.Server;
 	}
 
-	handleMessageChat(params)
+	skipExecute(params, extra = {})
 	{
+		if (!extra.optionImportant)
+		{
+			if (this.option.skip)
+			{
+				console.info('Pull: command skipped while loading messages', params);
+				return true;
+			}
+
+			if (!this.option.handlingDialog)
+			{
+				return false;
+			}
+		}
+
+		if (typeof params.chatId !== 'undefined' || typeof params.dialogId !== 'undefined' )
+		{
+			if (
+				typeof params.chatId !== 'undefined'
+				&& parseInt(params.chatId) === parseInt(this.option.handlingDialog.chatId)
+			)
+			{
+				return false;
+			}
+
+			if (
+				typeof params.dialogId !== 'undefined'
+				&& params.dialogId.toString() === this.option.handlingDialog.dialogId.toString()
+			)
+			{
+				return false;
+			}
+
+			return true;
+		}
+
+		return false;
+	}
+
+	handleMessage(params, extra)
+	{
+		this.handleMessageAdd(params, extra);
+	}
+
+	handleMessageChat(params, extra)
+	{
+		this.handleMessageAdd(params, extra);
+	}
+
+	handleMessageAdd(params, extra)
+	{
+		if (this.skipExecute(params, extra))
+		{
+			return false;
+		}
+
 		if (params.chat && params.chat[params.chatId])
 		{
 			this.store.dispatch('dialogues/update', {
-				dialogId: 'chat'+params.chatId,
+				dialogId: params.dialogId,
 				fields: params.chat[params.chatId]
 			});
 		}
@@ -56,9 +126,30 @@ class ImPullCommandHandler
 
 		if (params.files)
 		{
-			this.store.dispatch('files/set', this.controller.prepareFilesBeforeSave(
-				 VuexBuilderModel.convertToArray(params.files)
-			));
+			let files = VuexBuilderModel.convertToArray(params.files);
+			files.forEach(file =>
+			{
+				file = this.controller.prepareFilesBeforeSave(file);
+				if (
+					files.length === 1
+					&& params.message.templateFileId
+					&& this.store.state.files.index[params.chatId]
+					&& this.store.state.files.index[params.chatId][params.message.templateFileId]
+				)
+				{
+					this.store.dispatch('files/update', {
+						id: params.message.templateFileId,
+						chatId: params.chatId,
+						fields: file
+					}).then(() => {
+						this.controller.emit(EventType.dialog.scrollToBottom, {cancelIfScrollChange: true});
+					});
+				}
+				else
+				{
+					this.store.dispatch('files/set', file);
+				}
+			});
 		}
 
 		let collection = this.store.state.messages.collection[params.chatId];
@@ -68,61 +159,59 @@ class ImPullCommandHandler
 		}
 
 		let update = false;
-		if (params.message.tempId && collection.length > 0)
+		if (params.message.templateId && collection.length > 0)
 		{
 			for (let index = collection.length-1; index >= 0; index--)
 			{
-				if (collection[index].id == params.message.tempId)
+				if (collection[index].id === params.message.templateId)
 				{
 					update = true;
 					break;
 				}
 			}
 		}
+
 		if (update)
 		{
 			this.store.dispatch('messages/update', {
-				id: params.message.tempId,
-				chatId: params.message.chatId,
-				fields: params.message
+				id: params.message.templateId,
+				chatId: params.chatId,
+				fields: {
+					push: false,
+					...params.message,
+					sending: false,
+					error: false,
+				}
+			}).then(() => {
+				this.controller.emit(EventType.dialog.scrollToBottom, {cancelIfScrollChange: params.message.senderId !== this.controller.getUserId()});
 			});
 		}
 		else if (this.controller.isUnreadMessagesLoaded())
 		{
-			let unreadCountInCollection = 0;
-			if (collection.length > 0)
+			if (this.controller.getChatId() === params.chatId)
 			{
-				collection.forEach(element => element.unread? unreadCountInCollection++: 0);
+				this.store.commit('application/increaseDialogExtraCount');
 			}
 
-			if (unreadCountInCollection > 0)
-			{
-				this.store.commit('application/set', {dialog: {
-					messageLimit: this.controller.getRequestMessageLimit() + unreadCountInCollection
-				}});
-			}
-			else if (this.controller.getMessageLimit() != this.controller.getRequestMessageLimit())
-			{
-				this.store.commit('application/set', {dialog: {
-					messageLimit: this.controller.getRequestMessageLimit()
-				}});
-			}
-
-			this.store.dispatch('messages/set', {...params.message, unread: true});
+			this.store.dispatch('messages/setAfter', {
+				push: false,
+				...params.message,
+				unread: true
+			});
 		}
 
 		this.controller.stopOpponentWriting({
-			dialogId: 'chat'+params.message.chatId,
+			dialogId: params.dialogId,
 			userId: params.message.senderId
 		});
 
-		if (params.message.senderId == this.controller.getUserId())
+		if (params.message.senderId === this.controller.getUserId())
 		{
 			this.store.dispatch('messages/readMessages', {
-				chatId: params.message.chatId
+				chatId: params.chatId
 			}).then(result => {
 				this.store.dispatch('dialogues/update', {
-					dialogId: 'chat'+params.message.chatId,
+					dialogId: params.dialogId,
 					fields: {
 						counter: 0,
 					}
@@ -132,7 +221,7 @@ class ImPullCommandHandler
 		else
 		{
 			this.store.dispatch('dialogues/increaseCounter', {
-				dialogId: 'chat'+params.message.chatId,
+				dialogId: params.dialogId,
 				count: 1,
 			});
 		}
@@ -148,8 +237,39 @@ class ImPullCommandHandler
 		this.execMessageUpdateOrDelete(params, extra, command);
 	}
 
-	handleMessageDeleteComplete(params)
+	execMessageUpdateOrDelete(params, extra, command)
 	{
+		if (this.skipExecute(params, extra))
+		{
+			return false;
+		}
+
+		this.controller.stopOpponentWriting({
+			dialogId: params.dialogId,
+			userId: params.senderId
+		});
+
+		this.store.dispatch('messages/update', {
+			id: params.id,
+			chatId: params.chatId,
+			fields: {
+				text: command === "messageUpdate"? params.text: '',
+				textOriginal: command === "messageUpdate"? params.textOriginal: '',
+				params: params.params,
+				blink: true
+			}
+		}).then(() => {
+			this.controller.emit(EventType.dialog.scrollToBottom, {cancelIfScrollChange: true});
+		});
+	}
+
+	handleMessageDeleteComplete(params, extra)
+	{
+		if (this.skipExecute(params, extra))
+		{
+			return false;
+		}
+
 		this.store.dispatch('messages/delete', {
 			id: params.id,
 			chatId: params.chatId,
@@ -162,22 +282,68 @@ class ImPullCommandHandler
 		});
 	}
 
-	handleMessageParamsUpdate(params)
+	handleMessageLike(params, extra)
 	{
+		if (this.skipExecute(params, extra))
+		{
+			return false;
+		}
+
+		this.store.dispatch('messages/update', {
+			id: params.id,
+			chatId: params.chatId,
+			fields: {params: {LIKE: params.users}}
+		});
+	}
+
+	handleChatOwner(params, extra)
+	{
+		if (this.skipExecute(params, extra))
+		{
+			return false;
+		}
+
+		this.store.dispatch('dialogues/update', {
+			dialogId: params.dialogId,
+			fields: {
+				ownerId: params.userId,
+			}
+		});
+	}
+
+	handleMessageParamsUpdate(params, extra)
+	{
+		if (this.skipExecute(params, extra))
+		{
+			return false;
+		}
+
 		this.store.dispatch('messages/update', {
 			id: params.id,
 			chatId: params.chatId,
 			fields: {params: params.params}
+		}).then(() => {
+			this.controller.emit(EventType.dialog.scrollToBottom, {cancelIfScrollChange: true});
 		});
 	}
 
-	handleStartWriting(params)
+	handleStartWriting(params, extra)
 	{
+		if (this.skipExecute(params, extra))
+		{
+			return false;
+		}
+
 		this.controller.startOpponentWriting(params);
 	}
 
-	handleReadMessageChat(params)
+	handleReadMessage(params, extra)
 	{
+		if (this.skipExecute(params, extra))
+		{
+			return false;
+		}
+
 		this.store.dispatch('messages/readMessages', {
 			chatId: params.chatId,
 			readId: params.lastId
@@ -191,22 +357,73 @@ class ImPullCommandHandler
 		});
 	}
 
-	execMessageUpdateOrDelete(params, extra, command)
+	handleReadMessageChat(params, extra)
 	{
-		this.store.dispatch('messages/update', {
-			id: params.id,
-			chatId: params.chatId,
-			fields: {
-				text: command == "messageUpdate"? params.text: '',
-				textOriginal: command == "messageUpdate"? params.textOriginal: '',
-				params: params.params,
-				blink: true
-			}
-		});
+		this.handleReadMessage(params, extra);
+	}
 
-		this.controller.stopOpponentWriting({
+	handleReadMessageOpponent(params, extra)
+	{
+		this.execReadMessageOpponent(params, extra);
+	}
+
+	handleReadMessageChatOpponent(params, extra)
+	{
+		this.execReadMessageOpponent(params, extra);
+	}
+
+	execReadMessageOpponent(params, extra)
+	{
+		if (this.skipExecute(params, extra))
+		{
+			return false;
+		}
+
+		this.store.dispatch('dialogues/updateReaded', {
 			dialogId: params.dialogId,
-			userId: params.senderId
+			userId: params.userId,
+			userName: params.userName,
+			messageId: params.lastId,
+			date: params.date,
+			action: true
+		});
+	}
+
+	handleUnreadMessageOpponent(params, extra)
+	{
+		this.execUnreadMessageOpponent(params, extra);
+	}
+
+	handleUnreadMessageChatOpponent(params, extra)
+	{
+		this.execUnreadMessageOpponent(params, extra);
+	}
+
+	execUnreadMessageOpponent(params, extra)
+	{
+		if (this.skipExecute(params, extra))
+		{
+			return false;
+		}
+
+		this.store.dispatch('dialogues/updateReaded', {
+			dialogId: params.dialogId,
+			userId: params.userId,
+			action: false
+		});
+	}
+
+	handleFileUpload(params, extra)
+	{
+		if (this.skipExecute(params, extra))
+		{
+			return false;
+		}
+
+		this.store.dispatch('files/set', this.controller.prepareFilesBeforeSave(
+			 VuexBuilderModel.convertToArray({file: params.fileParams})
+		)).then(() => {
+			this.controller.emit(EventType.dialog.scrollToBottom, {cancelIfScrollChange: true});
 		});
 	}
 }
