+++ modules/main/lib/ui/viewer/filepreviewtable.php
@@ -83,8 +83,13 @@ final class FilePreviewTable extends DataManager
 
 		foreach ($files as $file)
 		{
-			self::deleteContent($file);
-			self::delete($file['ID']);
+			$keepImage = isset($file['PREVIEW_IMAGE_ID']);
+
+			self::deleteContent($file, $keepImage);
+			if (!$keepImage)
+			{
+				self::delete($file['ID']);
+			}
 		}
 	}
 
@@ -112,7 +117,7 @@ final class FilePreviewTable extends DataManager
 		self::deleteContent($file);
 	}
 
-	protected static function deleteContent(array $file)
+	protected static function deleteContent(array $file, $keepImage = false)
 	{
 		if (isset(self::$alreadyDeleted[$file['ID']]))
 		{
@@ -122,7 +127,10 @@ final class FilePreviewTable extends DataManager
 		self::$alreadyDeleted[$file['ID']] = true;
 
 		\CFile::delete($file['PREVIEW_ID']);
-		\CFile::delete($file['PREVIEW_IMAGE_ID']);
+		if (!$keepImage)
+		{
+			\CFile::delete($file['PREVIEW_IMAGE_ID']);
+		}
 	}
 
 	/**
